Refactor: address count 추가 및 관련 코드 리팩토링 ( #24 )
- 신고글 접수
- 신고글 삭제
- 주변 신고 위치 목록 조회
- MappingJackson2HttpMessageConverter

+++ src/main/java/somsomcore/zipcheck/domain/report/dto/report/ReportResponseDTO.java
@@ -69,7 +69,7 @@ public class ReportResponseDTO {
         private double latitude;
         private double longitude;
         private String address;
-        private int reportCount;
+        private Long reportCount;
     }
 
     // 주변 신고 위치 목록
